Electronic attendance system by using RFID system
Educational institutions’ administrators are concerned about student
irregular attendance. Checking students' attendance is one of the
important issues for universities, because many universities evaluate
students’ attendance while giving the final grade, professors consider
their total number of appearances on classes during the whole semester.
This brings to the idea of having some tool to control students’
attendance. However, Current manual way
of taking student attendance is not an efficient way since there will be
spent much of time for calling students names and putting marks like
"presence" or "absence" if the class
is a lecture class, and in this class at least 5 groups are presented.
Moreover, some students may call his/her friend as "presence" even
though this student is currently absent. After
thinking all these issues, author of the following research paper
decided to create a system that makes easier to check students’
attendance automatically. The system is based on
RFID technology, and in this paper, details of this system are
presented. The system can be easily accessed by the lecturers via the
web and most importantly, the reports can be
generated in real-time processing, thus, providing valuable information
about the students’.